functions/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.14.0
PyPDF2==3.0.1
python-docx==1.1.0
google-cloud-logging==3.8.0